pyproject.toml
[project]
name = "resumer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "pillow>=11.2.1",
    "pip>=25.1.1",
    "pydantic>=2.11.5",
    "uvicorn>=0.34.3",
]
